refactor: subfolder for Error Views
